fix(presidio): avoid reportArgumentType on tool call helper; move tests to covered suite
- type the _with_tool_call_arguments parameter as Any: the TypedDict input
  union (ChatCompletionToolCallChunk | ChatCompletionMessageToolCall) is not
  assignable to dict[str, Any], which breached the basedpyright
  reportArgumentType budget; runtime narrowing already happens in
  _get_tool_call_arguments
- move the four new tests from tests/guardrails_tests/ to
  tests/test_litellm/proxy/guardrails/guardrail_hooks/test_presidio.py so the
  coverage job exercises the new branch (codecov/patch)
