+++ tests/guardrails_tests/test_presidio_pii.py
@@ -587,138 +587,3 @@ async def test_presidio_language_configuration_with_per_request_override():
     # Verify the default language (German) is used
     assert analyze_request_default["language"] == "de"
     assert analyze_request_default["text"] == test_text
-
-
-### TOOL CALL ARGUMENT UNMASKING (issue #31950) ###
-
-TOOL_CALL_PII_TOKENS = {"<PERSON_1>": "Jane", "<PHONE_NUMBER_2>": "555-123-4567"}
-MASKED_TOOL_ARGUMENTS = (
-    '{"first_name": "<PERSON_1>", "last_name": "Doe", "phone": "<PHONE_NUMBER_2>"}'
-)
-UNMASKED_TOOL_ARGUMENTS = (
-    '{"first_name": "Jane", "last_name": "Doe", "phone": "555-123-4567"}'
-)
-
-
-def _masked_tool_call_dict() -> dict:
-    return {
-        "id": "call_1",
-        "type": "function",
-        "function": {"name": "lookup_reservation", "arguments": MASKED_TOOL_ARGUMENTS},
-    }
-
-
-@pytest.mark.asyncio
-async def test_apply_guardrail_unmasks_tool_call_arguments():
-    """
-    apply_guardrail must unmask PII tokens inside tool call arguments on the
-    response path — the unified guardrail hook routes /chat/completions
-    responses through apply_guardrail, not async_post_call_success_hook.
-    """
-    pii_masking = _OPTIONAL_PresidioPIIMasking(mock_testing=True)
-
-    result = await pii_masking.apply_guardrail(
-        inputs={"texts": [], "tool_calls": [_masked_tool_call_dict()]},
-        request_data={"metadata": {"pii_tokens": TOOL_CALL_PII_TOKENS}},
-        input_type="response",
-    )
-
-    assert result["tool_calls"][0]["function"]["arguments"] == UNMASKED_TOOL_ARGUMENTS
-
-
-@pytest.mark.asyncio
-async def test_apply_guardrail_tool_call_unmask_ignores_malformed_entries():
-    """Malformed tool call entries must pass through untouched, without raising."""
-    pii_masking = _OPTIONAL_PresidioPIIMasking(mock_testing=True)
-    malformed = [
-        {"id": "call_1", "type": "function"},  # no function
-        {"id": "call_2", "type": "function", "function": {"name": "f"}},  # no arguments
-        {"id": "call_3", "type": "function", "function": {"name": "f", "arguments": {"a": 1}}},  # non-str arguments
-    ]
-
-    result = await pii_masking.apply_guardrail(
-        inputs={"texts": [], "tool_calls": [dict(tc) for tc in malformed]},
-        request_data={"metadata": {"pii_tokens": TOOL_CALL_PII_TOKENS}},
-        input_type="response",
-    )
-
-    assert result["tool_calls"] == malformed
-
-
-@pytest.mark.asyncio
-async def test_apply_guardrail_masks_tool_call_arguments_on_request():
-    """On the request path, tool call arguments must be masked like texts are."""
-    from unittest.mock import AsyncMock
-
-    pii_masking = _OPTIONAL_PresidioPIIMasking(mock_testing=True)
-    pii_masking.check_pii = AsyncMock(return_value=MASKED_TOOL_ARGUMENTS)
-
-    tool_call = {
-        "id": "call_1",
-        "type": "function",
-        "function": {"name": "lookup_reservation", "arguments": UNMASKED_TOOL_ARGUMENTS},
-    }
-    result = await pii_masking.apply_guardrail(
-        inputs={"texts": [], "tool_calls": [tool_call]},
-        request_data={},
-        input_type="request",
-    )
-
-    pii_masking.check_pii.assert_awaited_once()
-    assert pii_masking.check_pii.await_args.kwargs["text"] == UNMASKED_TOOL_ARGUMENTS
-    assert result["tool_calls"][0]["function"]["arguments"] == MASKED_TOOL_ARGUMENTS
-
-
-@pytest.mark.asyncio
-async def test_unified_post_call_unmasks_tool_call_arguments_end_to_end():
-    """
-    End-to-end repro of issue #31950: a /chat/completions response whose only
-    payload is a tool call (content=None) must come back with tool call
-    arguments unmasked after the unified guardrail output processing.
-    """
-    from litellm.llms.openai.chat.guardrail_translation.handler import (
-        OpenAIChatCompletionsHandler,
-    )
-    from litellm.types.utils import ModelResponse
-
-    pii_masking = _OPTIONAL_PresidioPIIMasking(mock_testing=True)
-
-    response = ModelResponse(
-        **{
-            "id": "chatcmpl-1",
-            "created": 1,
-            "model": "gpt-5.4",
-            "object": "chat.completion",
-            "choices": [
-                {
-                    "finish_reason": "tool_calls",
-                    "index": 0,
-                    "message": {
-                        "role": "assistant",
-                        "content": None,
-                        "tool_calls": [
-                            {
-                                "id": "call_1",
-                                "type": "function",
-                                "function": {
-                                    "name": "lookup_reservation",
-                                    "arguments": MASKED_TOOL_ARGUMENTS,
-                                },
-                            }
-                        ],
-                    },
-                }
-            ],
-            "usage": {"prompt_tokens": 1, "completion_tokens": 1, "total_tokens": 2},
-        }
-    )
-
-    handler = OpenAIChatCompletionsHandler()
-    processed = await handler.process_output_response(
-        response=response,
-        guardrail_to_apply=pii_masking,
-        request_data={"metadata": {"pii_tokens": TOOL_CALL_PII_TOKENS}},
-    )
-
-    tool_call = processed.choices[0].message.tool_calls[0]
-    assert tool_call.function.arguments == UNMASKED_TOOL_ARGUMENTS
